gcs/pyproject.toml
[tool.poetry]
authors = ["Mert Kayhan <mk@fsncapital.com>"]
description = ""
name = "gcs"
version = "0.1.0"

[tool.poetry.dependencies]
google-cloud = "^0.34.0"
google-cloud-storage = "^2.3.0"
python = "^3.8"
utils = {url = "https://github.com/fsn-capital/pycommon/releases/download/test/utils-0.1.0.tar.gz"}

[tool.poetry.dev-dependencies]
coverage = "^6.4.2"
pytest = "^5.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
